arrange implementation of class GridGraph and Group

+++ libs/odpGridGraphs/src/fillEmptyPortRandomly.cpp
@@ -12,16 +12,16 @@ void FillEmptyPortRandomly::operator() (GridGraph& graph, std::mt19937& random) 
     int* nearNodes;
 
     vector<int> nodeSequence(graph.numNode());
-    for(int i = 0; i < nodeSequence.size(); ++i) nodeSequence[i] = i;
+    for(size_t i = 0; i < nodeSequence.size(); ++i) nodeSequence[i] = i;
     std::shuffle(nodeSequence.begin(), nodeSequence.end(), random);
 
-    for(int i = 0; i < nodeSequence.size(); ++i) {
+    for(size_t i = 0; i < nodeSequence.size(); ++i) {
         int nodeA = nodeSequence[i];
         if(graph.fullDegree(nodeA)) continue;
         
         vector<int> nearNodes = grid.selectConnectableNodes(nodeA);
         std::shuffle(nearNodes.begin(), nearNodes.end(), random);
-        for(int j = 0; j < nearNodes.size(); ++j) {
+        for(size_t j = 0; j < nearNodes.size(); ++j) {
             int nodeB = nearNodes[j];
             if(nodeA == nodeB)                  continue;
             if(graph.fullDegree(nodeB))         continue;
